refactor(team): keep task queries out of broker_tasks.go
